SetPixels now takes an ImageData offset for the x,y dirty rect placement. Added Vec2.setTo.
Moved the imageData index look-up to GetIndex and GetIndexFast, the later avoiding the bounds safety checks of the other. Updates GetPixel etc to use this new function.

Finished ProcessPixels and some tests for it. It will take an imageData and a region and pass each pixel to your own callback. The color that comes back from the callback is then set into the imageData. Useful for non-shader image manipulation.