Added Canvas functions for BeginPath, ClosePath, Save, Restore and Pattern.
Worked a lot on the ShapeFill class, which accompanies a Shape object. You can now stack multiple fills (or strokes) onto a single Shape and they'll be applied in the order specified.

Removed ShapeGradient and renamed it ShapeFill - adding in lots of helper functions to make the one class deal with all varieties of fill (solid, patter, linear and radial). Updated lots of examples to handle the new format too.