`Loader.getLoadedFilesByType` is a new method that allows you to get all of the loaded files of a specific type. All of the Loader file types have had new consts added to make them more secure. You can now for example do `getLoadedFilesByType(ImageFile.TYPE)` to get just the loaded images. There is an optional `group` argument as well.
The Cache has been recoded to be a pure object, rather than a class.

The Nano Game has been updated to include a default State, a working main loop and has the loader hooked up.

Added an Image parser to the Texture Frame Parsers. This will create a Frame object for a single image (needed by the Cache). Made the ImageFile loader set the image key as its `name` property. This property exists natively on the Image objects and is unset, so we may as well use it.